docs: Add 'design.adoc' to project documentation
A new file named 'design.adoc' has been created and included in 'base.adoc'. This new file contains the documentation for our design system using Storybook. It descibes how to access and use Storybook and the advantages it provides in terms of building and testing UI components independently. It also details how Storybook enhances team collaboration by acting as a central point of reference for all.