refactor: Update FDHorizontalMenu and auth.store components
This commit refactors the FDHorizontalMenu component by optimizing useAuthStore and useThemeStore hooks and their respective functions. The getUsername function was also added to the auth.store.ts file to provide an efficient way to retrieve the username. This refactor not only simplifies the management of authentication and theme state but also enhances the readability of the code.