refactor: update NotesList layout
New icon entries for DOCUMENT and DOCUMENT_FILLED are added in the FDIcon file. Several changes have been made in the layout and styling of the NotesList component. These include the use of the newly added icons, replacement of simple div elements with the new NotesListItem component which takes notes as props, the increase of width, and change of background color for better visibility. A new css file is also imported for extended styling options.