refactor: replace tailwind design system with material-design 3
The tailwind design system will be removed with the design system of google, to streamline development. The components FDButton, FDQuizCard, FDSwitch, NotesListItem, QuestionCardBase, QuestionCardBlank, RepeatQuestionnaireCard and various other components were removed from the codebase. Moreover, a small currency tweak was made to FDIcon and additional features were added to package-lock.json. This cleanup operation was necessary to streamline the code and remove unused or redundant components. We also updated some dependencies to their latest versions for better performance and compatibility.