refactor: decorators in resource.controller and api.decorator
Revised the usage and return of decorators in resource.controller.ts and api.decorator.ts. Moved the securityType to be in the same decorator as the schema and responseDescription, and applied the httpCode decorator to the FoxdeckApiResponse. These changes simplify the decorator application, and create more concise and organized usage in the code.