chore: API method to get User Resources and improved Select Query
Added a new REST API method in the resource controller to fetch all resources associated with a user. Also, enhanced the database select method in SQLite provider service to perform more complex queries. This leads to an improved method for gathering resources in the application by providing additional ways to retrieve data.