refactor: SQLite provider service and resource store
This commit updates the SQLite provider service and resource store. Specifically, the condition for 'singleSelect' was changed to check for a boolean value and unnecessary code handling one-item arrays was removed in the SQLite service. Additionally, an empty array was assigned as the intial state for 'resources' and a console log was added in the resource store. Lastly, a small syntax change was made in the 'getAllResourcesByUserId' method in the resource service.