Replace AppTextField with new AppSearch component
The AppTextField component in the HomeView module has been replaced with the new AppSearch component. This change includes the creation of the AppSearch module, which is now responsible for handling search functionality, providing a dedicated and more flexible venue for implementing search features.

Currently, this component is work in progress and only shows the expected ui without logic. I will work further into this.