Add note search by name feature to API
Added a new function to the resource controller to search notes by name, and also updated the API documentation accordingly. The method uses a mix between the soundex algorithm for comparing similarity between the provided name and existing notes and a simple like query. Also adjusted routing and client-side service for this new functionality.

We also transformed the response of the Resources from 'title case' to 'lowercase'.